+++ src/Event/mt_evtLoadSave.c
@@ -47,7 +47,7 @@ void FUN_003905f0(int param_1,int param_2);
 void FUN_00390920(int param_1,int param_2);
 void FUN_00390c90(int param_1,int param_2);
 void FUN_00390ef0(int param_1,int param_2);
-void FUN_00391080(u32 param_1,u32 param_2);
+void FUN_00391080(u64 param_1,u64 param_2);
 u32 FUN_003912f0(u32 param_1,u32 param_2);
 
 
@@ -3213,7 +3213,7 @@ void FUN_00390ef0(int param_1,int param_2)
 // FUN_00391080 NONMATCHING
 
 
-void FUN_00391080(u32 param_1,u32 param_2)
+void FUN_00391080(u64 param_1,u64 param_2)
